+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Added
+- multi-term filtering to /api/v1/items
+
 ## [2.72.1] - 2023-02-10
 ### Added
 - `disk` property to Imports